# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant email SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with username or email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Sign-in temporarily limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a tenant account and send email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Verification instructions sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Application email is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/verify-email:
    post:
      operationId: verifyRegistrationEmail
      tags: [auth]
      summary: Verify a registration email with a one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified and session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Send a password reset link without revealing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Generic response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a one-time reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed; other sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid password or current password mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [SUPERADMIN, USER] }
        timezone: { type: string }
        active: { type: boolean }
        emailVerified: { type: boolean }
        mustChangeCredentials: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, username, firstName, lastName, email, role, timezone, active, emailVerified, mustChangeCredentials, createdAt]
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
      required: [user]
    LoginInput:
      type: object
      properties:
        identifier: { type: string, minLength: 1, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 256 }
      required: [identifier, password]
    RegisterInput:
      type: object
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 256 }
      required: [firstName, lastName, email, password]
    VerifyEmailInput:
      type: object
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 6, maxLength: 6 }
      required: [email, code]
    ForgotPasswordInput:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        token: { type: string, minLength: 32, maxLength: 256 }
        password: { type: string, minLength: 8, maxLength: 256 }
      required: [token, password]
    ChangePasswordInput:
      type: object
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 256 }
        newPassword: { type: string, minLength: 8, maxLength: 256 }
      required: [currentPassword, newPassword]
...