# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant email SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  username: string;
  firstName: string;
  lastName: string;
  email: string;
  role: AuthUserRole;
  timezone: string;
  active: boolean;
  emailVerified: boolean;
  mustChangeCredentials: boolean;
  createdAt: Date;
}