# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant email SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 256
     */
  password: string;
}