# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant email SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}